test: move TAS tests to tas/baseline and tas/extended subfolders
