+++ test/e2e/tas/baseline/job_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package tasbaseline
+package baseline
 
 import (
 	"fmt"
