+++ test/e2e/tas/extended/rayjob_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package tasextended
+package extended
 
 import (
 	"github.com/onsi/ginkgo/v2"
